Move dev stubs and such to their own files.
