+++ gulpfile.js
@@ -5,6 +5,7 @@ const babel = require('gulp-babel');
 const colors = require('gulp-util').colors;
 const log = require('gulp-util').log;
 const through2 = require('through2');
+const replace = require('replacestream');
 const gulpif = require('gulp-if');
 const uglify = require('gulp-uglify');
 const concat = require('gulp-concat');
@@ -53,76 +54,6 @@ function getPort() {
   const { port } = server.address();
   server.close();
   return port;
-}
-function dev(done) {
-  const port = getPort();
-  const publicPath = `https://localhost:${port}/`;
-  const address = `${publicPath}dev.js`;
-
-  webpackConfig.entry.unshift(
-    `${require.resolve('webpack-dev-server/client')}?${publicPath}`,
-    'webpack/hot/dev-server');
-  webpackConfig.entry.pop();
-  webpackConfig.entry.push('./hotReloader.js');
-  webpackConfig.output.publicPath = publicPath;
-  webpackConfig.output.filename = 'dev.js';
-  webpackConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
-  const compiler = webpack(webpackConfig);
-  const server = new DevServer(compiler, {
-    port,
-    publicPath,
-    contentBase: webpackConfig.output.path,
-    hot: true,
-    https: true,
-    disableHostCheck: true,
-    headers: {
-      'Access-Control-Allow-Origin': '*',
-    },
-    setup(app) {
-      app.get('/ok.html', (req, res) => {
-        res.setHeader('content-type', 'text/html');
-        res.send(`
-          <!DOCTYPE html>
-          <body style="font: 200% sans-serif; max-width: 600px; margin: auto">
-            <h1>All good!</h1>
-            <p>You can now close this tab.</p>
-            <script>
-              try { window.opener.postMessage('ok') } catch (e) {}
-              window.close()
-            </script>
-          </body>
-        `);
-      });
-    },
-  });
-
-  server.listen(port, () => {
-    gulp.series(buildLoader, function writeDevStub(done) {
-      fs.writeFile('build/source.js', `
-        requirejs.config({
-          paths: { 'extplug/__internalExtPlug__': ${JSON.stringify(address)}.replace(/\.js$/, '') }
-        });
-        function tryAccess() {
-          return window.jQuery.get(${JSON.stringify(publicPath)});
-        }
-        tryAccess().fail(function () {
-          require('plug/core/Events').trigger('notify', 'icon-chat-system',
-            'Error: Could not load the ExtPlug dev mode extension. This is probably because its certificate is not whitelisted. ' +
-            'Please click here and whitelist the certificate.', function () {
-              window.open(${JSON.stringify(`${publicPath}ok.html`)});
-              var i = setInterval(function () {
-                tryAccess().then(function () {
-                  clearInterval(i);
-                  window.location.reload();
-                }).fail(function () {
-                  console.log('waiting for approval...');
-                });
-              }, 1500);
-            });
-        });
-      `, done);
-    }, concatSource, wrapBuiltSourceInLoader, buildExtensions)(done);
-  });
 }
 
 function buildLoaderTransform() {
@@ -250,6 +181,48 @@ const buildExtensions = gulp.parallel(
   buildFirefoxExtension,
   buildUserscript
 );
+
+function dev(done) {
+  const port = getPort();
+  const publicPath = `https://localhost:${port}/`;
+  const address = `${publicPath}dev.js`;
+
+  webpackConfig.entry.unshift(
+    `${require.resolve('webpack-dev-server/client')}?${publicPath}`,
+    'webpack/hot/dev-server');
+  webpackConfig.entry.pop();
+  webpackConfig.entry.push('./hotReloader.js');
+  webpackConfig.output.publicPath = publicPath;
+  webpackConfig.output.filename = 'dev.js';
+  webpackConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
+  const compiler = webpack(webpackConfig);
+  const server = new DevServer(compiler, {
+    port,
+    publicPath,
+    contentBase: webpackConfig.output.path,
+    hot: true,
+    https: true,
+    disableHostCheck: true,
+    headers: {
+      'Access-Control-Allow-Origin': '*',
+    },
+    setup(app) {
+      app.get('/ok.html', (req, res) => {
+        res.setHeader('content-type', 'text/html');
+        fs.createReadStream('src/devCertificateHelper.html').pipe(res);
+      });
+    },
+  });
+
+  server.listen(port, () => {
+    gulp.series(buildLoader, function writeDevStub(done) {
+      return fs.createReadStream('src/devStub.js')
+        .pipe(replace('PUBLIC_PATH', JSON.stringify(publicPath)))
+        .pipe(replace('ADDRESS', JSON.stringify(address)))
+        .pipe(fs.createWriteStream('build/source.js'));
+    }, concatSource, wrapBuiltSourceInLoader, buildExtensions)(done);
+  });
+}
 
 exports.default = gulp.series(
   cleanBuild,
