+++ package.json
@@ -51,6 +51,7 @@
     "postcss-cssnext": "^2.8.0",
     "postcss-import": "^10.0.0",
     "postcss-loader": "^2.0.0",
+    "replacestream": "^4.0.2",
     "requirejs": "^2.1.20",
     "through2": "^2.0.1",
     "url-loader": "^0.5.7",
